+++ Ogma3/Data/DTOs/CommentFromApiDTO.cs
@@ -3,6 +3,6 @@ namespace Ogma3.Data.DTOs
     public class CommentFromApiDTO
     {
         public string Body { get; set; }
-        public int Thread { get; set; }
+        public long Thread { get; set; }
     }
 }
