+++ Ogma3/Data/DTOs/VoteFromApiDTO.cs
@@ -2,6 +2,6 @@ namespace Ogma3.Data.DTOs
 {
     public class VoteFromApiDTO
     {
-        public int VotePool { get; set; }
+        public long VotePool { get; set; }
     }
 }
